feat: better credentialess errors (#163)
